money/requirements.txt
Flask==1.1.1
Jinja2==2.10.3
pysummarize==0.6.0
python-dotenv==0.10.5
python-rake==1.4.5
